moved unused external setters to private added cast to int for size on vector fixed closed state update for drawer

+++ Shared/messages/MessageFeedsController.cpp
@@ -112,7 +112,7 @@ void MessageFeedsController::findEntities(const QString& entityIdText)
 
 void MessageFeedsController::selectEntity(int index)
 {
-  if (index < 0 || index >= m_entityIds.size())
+  if (index < 0 || index >= static_cast<int>(m_entityIds.size()))
     return;
 
   emit entitySelected(m_entityIds[index], selectedFeed());
